[22-10-07] 그룹, 인증샷 수정시 이미지 삭제 오류 수정

+++ src/main/java/com/havit/finalbe/service/CertifyService.java
@@ -147,7 +147,9 @@ public class CertifyService {
         }
 
         Long originImage = certify.getImageId();
-        imageService.deleteImage(originImage);
+        if (!originImage.equals(certifyRequestDto.getImageId())) {
+            imageService.deleteImage(originImage);
+        }
         certify.update(certifyRequestDto);
 
         return CertifyResponseDto.builder()
